Moved third-party library files to separate directories.
