+++ src/excrement.py
@@ -1,6 +1,6 @@
 import random
 from drawableEntity import DrawableEntity
-from entity import Entity
+from pyenvlib.entity import Entity
 
 
 # @author Daniel McCoy Stephenson
