+++ src/excreteActionHandler.py
@@ -1,9 +1,9 @@
 import random
-from entity import Entity
+from pyenvlib.entity import Entity
 
 from excrement import Excrement
-from grid import Grid
-from location import Location
+from pyenvlib.grid import Grid
+from pyenvlib.location import Location
 
 
 # @author Daniel McCoy Stephenson
